requirements.txt
asgiref==3.4.1
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.46.0
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-imagekit==4.1.0
django-social-share==0.3.0
django-sslserver==0.22
django-widget-tweaks==1.4.9
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
path.py==8.1.2
pilkit==2.0
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
Unidecode==0.4.19
urllib3==1.26.7
whitenoise==5.3.0
